+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [24.1.2] 2020-05-04
+### Fixed
+- Fixed contained valueset reference for R4 (from using url to using #vs-id),
+  but continue to support url-based reference only for the purpose of
+  backward-compatibility with previous LForms versions.
+
 ## [24.1.1] 2020-04-28
 ### Changed
 - Updated the fhirpath dependency to 2.1.3